requirements.txt
robyn==0.64.2
python-dotenv==1.0.1
asyncmy==0.2.10
greenlet==3.1.1
aiomysql==0.2.0
pydantic==2.10.4
tortoise-orm==0.22.2
